chore: add feedback form in dashboard
