+++ src/blocks/plugins/registerPlugin.js
@@ -25,7 +25,7 @@ import './menu-icons/index.js';
 import './sticky/index.js';
 import './dynamic-content/index.js';
 import './welcome-guide/index.js';
-import './feedback';
+import './feedback/index.tsx';
 
 const icon = <Icon icon={ otterIcon } />;
